remove extension-store for now
