--- src/renderer/components/+extensions/extensions.route.ts
@@ -1,11 +1,0 @@
-import { RouteProps } from "react-router";
-import { buildURL } from "../../navigation";
-
-export const extensionsRoute: RouteProps = {
-  path: "/extensions"
-}
-
-export interface IExtensionsRouteParams {
-}
-
-export const extensionsURL = buildURL<IExtensionsRouteParams>(extensionsRoute.path);
