--- src/renderer/components/+extensions/extensions.scss
@@ -1,4 +1,0 @@
-.Extensions {
-  $spacing: $padding * 2;
-  padding: $spacing;
-}
