--- src/renderer/components/+extensions/index.ts
@@ -1,2 +1,0 @@
-export * from "./extensions.route"
-export * from "./extensions"
